stock_release_channel: prefetch data before release
while `release_available_to_promise` is being executed, release_channel_id is set
on stock pickings, which triggers an inverse method setting
expected_date on the whole chain of moves.

For each picking, it would execute the same sql query, which is takes a
lot of time mapping `stock_move_move_rel` with `stock_move.company_id`.

In order to fix that, fetch data in the cache before calling super() in
the override of release_available_to_promise().